applied patch from Erik de Castro Lopo bug filed by Dan (sweep at sasami.anime.net) on sweep-users:
> i confirmed sweep 0.8.2 and libsamplerate-0.0.15... and yes I am still
> having problems.
>
> http://bani.anime.net/hit1_4.wav
>
> mix down to mono -> normalize ->
>    best sinc interpolator resample to 22050 -> hang

Yep, I can confirm this.

This is also unrelated to the "mix down to mono" operation. If you first
mix down to mono, save the file and then work on the mono file from
there on, there is still a problem. The strange thing is that the problem
doesn't seem to happen with stereo files.

I also tried taking the mono version of the file and resampling it to
22050Hz using the command line program sndfile-resample (one of the
example programs in the libsamplrate package) where it works correctly.

Finally, when I resample the mono file in Sweep, I notice that when it
hangs, it starts chewing up CPU power suggesting that its caught in an
infinite loop.

Erik


git-svn-id: http://svn.metadecks.org/trunk/sweep@141 e2d53364-18fc-0310-981e-b60a84739af0